+++ .github/workflows/tests_linux_x86_mpi.yml
@@ -122,10 +122,11 @@ jobs:
           cmake --build ./Build
           cd ./Build
           mkdir -p ./tests/results
+          for file in *runner ; do ./$file --order lex --reporter junit --out ./tests/results/report_$file.xml; done;
           for file in *runner_mpi ; do /opt/mpi/${{ matrix.mpilib }}/bin/mpirun -np 2 ./$file --order lex --reporter junit --out ./tests/results/report_$file.xml; done;
           lcov --directory . -b ../pennylane_lightning/src --capture --output-file coverage.info
           lcov --remove coverage.info '/usr/*' --output-file coverage.info
-          mv coverage.info coverage-${{ github.job }}-lightning_gpu.info
+          mv coverage.info coverage-${{ github.job }}-lightning_gpu_cpp_${{ matrix.mpilib }}.info
       - name: Upload test results
         uses: actions/upload-artifact@v3
         if: always()
@@ -139,7 +140,7 @@ jobs:
         with:
           if-no-files-found: error
           name: ubuntu-codecov-results-cpp
-          path: ./Build/coverage-${{ github.job }}-lightning_gpu.info
+          path: ./Build/coverage-${{ github.job }}-lightning_gpu_cpp_${{ matrix.mpilib }}.info
 
   upload-to-codecov-linux-cpp:
     needs: [cpp_tests]
@@ -164,8 +165,5 @@ jobs:
         if: always()
         run: |
           rm -rf ${{ steps.setup_venv.outputs.venv_name }}
-          rm -rf *
-          rm -rf .git
-          rm -rf .gitignore
-          rm -rf .github
+          rm -rf * .git .gitignore .github
           pip cache purge
