+++ pennylane_lightning/core/src/simulators/lightning_gpu/MPIWorker.hpp
@@ -19,6 +19,7 @@
 #include <vector>
 
 #include "MPIManager.hpp"
+#include "MPI_helpers.hpp"
 #include "cuError.hpp"
 #include <cuda.h>
 #include <cuda_runtime.h>
@@ -32,79 +33,12 @@ using namespace Pennylane::LightningGPU;
 
 namespace Pennylane::LightningGPU::MPI {
 
-enum WireStatus { Default, Target, Control };
-
 inline size_t mebibyteToBytes(const size_t mebibytes) {
     return mebibytes * size_t{1024 * 1024};
 }
 
 inline double bytesToMebibytes(const size_t bytes) {
     return static_cast<double>(bytes) / (1024.0 * 1024.0);
-}
-
-/**
- * @brief Create wire pairs for bit index swap and transform all control and
- * target wires to local ones.
- *
- * @param numLocalQubits Number of local qubits.
- * @param numTotalQubits Number of total qubits.
- * @param ctrls Vector of control wires.
- * @param tgts Vector of target wires.
- * @return wirePairs Wire pairs to be passed to SV bit index swap worker.
- */
-inline std::vector<int2> createWirePairs(const int numLocalQubits,
-                                         const int numTotalQubits,
-                                         std::vector<int> &ctrls,
-                                         std::vector<int> &tgts,
-                                         std::vector<int> &statusWires) {
-    std::vector<int2> wirePairs;
-    int localbit = numLocalQubits - 1, globalbit = numLocalQubits;
-    while (localbit >= 0 && globalbit < numTotalQubits) {
-        if (statusWires[localbit] == WireStatus::Default &&
-            statusWires[globalbit] != WireStatus::Default) {
-            int2 wirepair = make_int2(localbit, globalbit);
-            wirePairs.push_back(wirepair);
-            if (statusWires[globalbit] == WireStatus::Control) {
-                for (size_t k = 0; k < ctrls.size(); k++) {
-                    if (ctrls[k] == globalbit) {
-                        ctrls[k] = localbit;
-                    }
-                }
-            } else {
-                for (size_t k = 0; k < tgts.size(); k++) {
-                    if (tgts[k] == globalbit) {
-                        tgts[k] = localbit;
-                    }
-                }
-            }
-            std::swap(statusWires[localbit], statusWires[globalbit]);
-        } else {
-            if (statusWires[localbit] != WireStatus::Default) {
-                localbit--;
-            }
-            if (statusWires[globalbit] == WireStatus::Default) {
-                globalbit++;
-            }
-        }
-    }
-    return wirePairs;
-}
-
-/**
- * @brief Create wire pairs for bit index swap and transform all target wires to
- * local ones.
- *
- * @param numLocalQubits Number of local qubits.
- * @param numTotalQubits Number of total qubits.
- * @param tgts Vector of target wires.
- * @return wirePairs Wire pairs to be passed to SV bit index swap worker.
- */
-inline std::vector<int2> createWirePairs(int numLocalQubits, int numTotalQubits,
-                                         std::vector<int> &tgts,
-                                         std::vector<int> &statusWires) {
-    std::vector<int> ctrls;
-    return createWirePairs(numLocalQubits, numTotalQubits, ctrls, tgts,
-                           statusWires);
 }
 
 /**
